chore: transform last tests from TypePostProcessing.cpp to type-name-print-tests suite
